Nest the Workload, Kotlin, Xml, Json and Teletype satellites
`WorkloadGrant`, `WorkloadHandle` and `WorkloadOpenable` become `Workload.*`;
`KotlinRuntime` becomes `Kotlin.Runtime`; `XmlReader`, `JsonReader` and
`TeletypeBuilder` become `Xml.Reader`, `Json.Reader` and `Teletype.Builder`.
Seven names leave the umbrella's exported surface.

Three of these shadowed a type they depend on, all fixed by qualifying the base
rather than abandoning the nesting: `Workload.Openable` extends
`aperture.Openable`, `Workload.Grant`'s members extend `aperture.Grant` (and one
`Granting[aperture.Grant.Read]`), and `Teletype.Builder` extends
`gossamer.Builder`.

`KotlinFacade` is deliberately **not** nested. `object Kotlin` refers to
xenophile's own `Facade` about ten times — in `Expr[Facade]`, `TypeRepr.of[Facade]`
and as a constructor — so `Kotlin.Facade` would put two different `Facade`s in one
file. That is a worse read than the compound name, and unlike a shadowed
*supertype* there is no single base to qualify.

`XmlSchema` stays whole under R2, and `KotlinDialect`/`KotlinInvoke` stay because
both are resolved reflectively by name.
