+++ lib/breviloquence/src/core/breviloquence.Cbor.scala
@@ -1680,7 +1680,7 @@ object Cbor extends Cbor2, Dynamic:
     // Only breviloquence's read path (`Cbor.parseDirect`) constructs readers,
     // so the exclusivity of the wrapped parser and the resolution scope of the
     // carried tactic are preserved by construction. The wrapped tactic travels
-    // as a neutral carrier (jacinta's `JsonReader` pattern): the field stays
+    // as a neutral carrier (jacinta's `Json.Reader` pattern): the field stays
     // pure, and each accessor reasserts the type at the rim — the audited
     // point.
     private[breviloquence] def apply(parser: Cbor.Parser, tactic: Tactic[Cbor.Error])
