+++ lib/embarcadero/src/containerd/embarcadero_containerd.scala
@@ -37,5 +37,5 @@ import aperture.*
 // Atomic modes selecting the workload grants, composed like aperture's `Read` and `Write`:
 // `container.open[Workload](Read & Launch & Signal)`. The default `Read` open creates the
 // container and its task but does not start it — an inspection scope.
-val Launch: Mode granting WorkloadGrant.Launch = new Mode { type Grants = WorkloadGrant.Launch }
-val Signal: Mode granting WorkloadGrant.Signal = new Mode { type Grants = WorkloadGrant.Signal }
+val Launch: Mode granting Workload.Grant.Launch = new Mode { type Grants = Workload.Grant.Launch }
+val Signal: Mode granting Workload.Grant.Signal = new Mode { type Grants = Workload.Grant.Signal }
