+++ lib/embarcadero/src/containerd/soundness_embarcadero_containerd.scala
@@ -40,5 +40,4 @@ export embarcadero.{AnyMessage, Containerd, Container, ContentDescriptor, Docker
     GetTaskResponse, ImageRecord, KillRequest, ListContainersRequest, ListContainersResponse,
     ListImagesRequest, ListImagesResponse, ListNamespacesRequest, ListNamespacesResponse,
     ListTasksRequest, ListTasksResponse, Mount, Namespace, Workload, ProcessStatus, Runtime,
-    StartRequest, StartResponse, VersionResponse, WaitRequest, WaitResponse, WorkloadGrant,
-    Launch, Signal, WorkloadHandle, WorkloadOpenable}
+    StartRequest, StartResponse, VersionResponse, WaitRequest, WaitResponse, Launch, Signal}
