--- lib/escapade/src/core/escapade.TeletypeBuilder.scala
@@ -1,90 +1,0 @@
-                                                                                                  /*
-┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━┓
-┃                                                                                                  ┃
-┃                                                   ╭───╮                                          ┃
-┃                                                   │   │                                          ┃
-┃                                                   │   │                                          ┃
-┃   ╭───────╮╭─────────╮╭───╮ ╭───╮╭───╮╌────╮╭────╌┤   │╭───╮╌────╮╭────────╮╭───────╮╭───────╮   ┃
-┃   │   ╭───╯│   ╭─╮   ││   │ │   ││   ╭─╮   ││   ╭─╮   ││   ╭─╮   ││   ╭─╮  ││   ╭───╯│   ╭───╯   ┃
-┃   │   ╰───╮│   │ │   ││   │ │   ││   │ │   ││   │ │   ││   │ │   ││   ╰─╯  ││   ╰───╮│   ╰───╮   ┃
-┃   ╰───╮   ││   │ │   ││   │ │   ││   │ │   ││   │ │   ││   │ │   ││   ╭────╯╰───╮   │╰───╮   │   ┃
-┃   ╭───╯   ││   ╰─╯   ││   ╰─╯   ││   │ │   ││   ╰─╯   ││   │ │   ││   ╰────╮╭───╯   │╭───╯   │   ┃
-┃   ╰───────╯╰─────────╯╰────╌╰───╯╰───╯ ╰───╯╰────╌╰───╯╰───╯ ╰───╯╰────────╯╰───────╯╰───────╯   ┃
-┃                                                                                                  ┃
-┃    Soundness, version 0.64.0.                                                                    ┃
-┃    © Copyright 2021-25 Jon Pretty, Propensive OÜ.                                                ┃
-┃                                                                                                  ┃
-┃    The primary distribution site is:                                                             ┃
-┃                                                                                                  ┃
-┃        https://soundness.dev/                                                                    ┃
-┃                                                                                                  ┃
-┃    Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file     ┃
-┃    except in compliance with the License. You may obtain a copy of the License at                ┃
-┃                                                                                                  ┃
-┃        https://www.apache.org/licenses/LICENSE-2.0                                               ┃
-┃                                                                                                  ┃
-┃    Unless required by applicable law or agreed to in writing,  software distributed under the    ┃
-┃    License is distributed on an "AS IS" BASIS,  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,    ┃
-┃    either express or implied. See the License for the specific language governing permissions    ┃
-┃    and limitations under the License.                                                            ┃
-┃                                                                                                  ┃
-┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━┛
-                                                                                                  */
-package escapade
-
-import scala.collection.mutable as scm
-
-import anticipation.*
-import gossamer.*
-import rudiments.*
-import vacuous.*
-
-class TeletypeBuilder(size: Optional[Int] = Unset) extends Builder[Teletype]:
-  private val builder: StringBuilder = StringBuilder()
-  private val styles: scm.ArrayBuffer[Long] = scm.ArrayBuffer.empty
-  private val hyperlinks: scm.HashMap[Int, Text] = scm.HashMap()
-  private val insertions: scm.TreeMap[Int, Text] = scm.TreeMap()
-
-  @scala.caps.unsafe.untrackedCaptures
-  private var offset: Int = 0
-
-  def length: Int = builder.length
-
-  protected def wipe(): Unit =
-    offset = 0
-    builder.clear()
-    styles.clear()
-    hyperlinks.clear()
-    insertions.clear()
-
-  protected def put(text: Teletype): Unit =
-    builder.append(text.plain.s)
-    var i = 0
-
-    while i < text.plain.length do
-      styles += text.styleAt(i)
-      i += 1
-
-    text.hyperlinks.stdlib.each: (k, v) => hyperlinks(k + offset) = v
-    text.insertions.each: (k, v) => insertions(k + offset) = v
-
-    offset += text.plain.length
-
-  protected def putChar(char: Char): Unit =
-    builder.append(char)
-    styles += 0L
-    offset += 1
-
-  protected def result(): Teletype =
-    styles += 0L
-
-    val plainText = builder.toString.tt
-    val denseStyles = Array.unsafeFrozen(styles.toArray)
-    val (newStyles, newBoundaries) = Teletype.compressIfBeneficial(plainText, denseStyles)
-
-    Teletype
-      ( plainText,
-        newStyles,
-        Map.of(hyperlinks.toMap),
-        insertions.to(TreeMap),
-        newBoundaries )
