+++ lib/escapade/src/core/soundness_escapade_core.scala
@@ -35,7 +35,7 @@ package soundness
 export
   escapade
   . { Ansi, Ansi2, Bg, Bold, CharSpan, Colorable, Conceal, csi, e, Escape, escapes, Fg, Imprintable,
-      Italic, Reverse, Ribbon, Strike, Stylize, Teletype, teletype, Teletypeable, TeletypeBuilder,
+      Italic, Reverse, Ribbon, Strike, Stylize, Teletype, teletype, Teletypeable,
       TextStyle, Underline, Hyperlink, TerminalEscapes }
 
 package displayableTypes:
