+++ lib/jacinta/src/staged/jacinta.Inlinable.scala
@@ -56,7 +56,7 @@ import prepositional.*
 // an earlier run than the expansion; same-run instances degrade to a spliced
 // runtime call through `Json.Field`.
 trait Inlinable extends Typeclass:
-  def parse(reader: Expr[JsonReader])(using Quotes, Type[Self]): Expr[Self]
+  def parse(reader: Expr[Json.Reader])(using Quotes, Type[Self]): Expr[Self]
 
   // What a field of this type yields when its key is absent from the object,
   // mirroring the runtime instances: an abort unless overridden.
@@ -85,7 +85,7 @@ object Inlinable:
     type Self = value
     private[jacinta] def delegate: value is Inlinable = delegate0
 
-    def parse(reader: Expr[JsonReader])(using Quotes, Type[value]): Expr[value] =
+    def parse(reader: Expr[Json.Reader])(using Quotes, Type[value]): Expr[value] =
       delegate0.parse(reader)
 
     override def absent(tactic: Expr[Tactic[Json.Error]])(using Quotes, Type[value])
@@ -99,14 +99,14 @@ object Inlinable:
   private[jacinta] final class ProductInlinable[product]() extends Inlinable:
     type Self = product
 
-    def parse(reader: Expr[JsonReader])(using Quotes, Type[product]): Expr[product] =
+    def parse(reader: Expr[Json.Reader])(using Quotes, Type[product]): Expr[product] =
       stagedInternal.productBody[product](reader)
 
   private[jacinta] final class IterableInlinable[element](element0: element is Inlinable)
   extends Inlinable:
     type Self = Iterable[element]
 
-    def parse(reader: Expr[JsonReader])(using Quotes, Type[Iterable[element]])
+    def parse(reader: Expr[Json.Reader])(using Quotes, Type[Iterable[element]])
     :   Expr[Iterable[element]] =
 
       stagedInternal.iterableBody[Iterable[element]](reader, element0)
@@ -118,37 +118,37 @@ object Inlinable:
 
   given int: (Int is Inlinable) = new Inlinable:
     type Self = Int
-    def parse(reader: Expr[JsonReader])(using Quotes, Type[Int]): Expr[Int] =
+    def parse(reader: Expr[Json.Reader])(using Quotes, Type[Int]): Expr[Int] =
       '{ $reader.long().toInt }
 
   given long: (Long is Inlinable) = new Inlinable:
     type Self = Long
-    def parse(reader: Expr[JsonReader])(using Quotes, Type[Long]): Expr[Long] =
+    def parse(reader: Expr[Json.Reader])(using Quotes, Type[Long]): Expr[Long] =
       '{ $reader.long() }
 
   given double: (Double is Inlinable) = new Inlinable:
     type Self = Double
-    def parse(reader: Expr[JsonReader])(using Quotes, Type[Double]): Expr[Double] =
+    def parse(reader: Expr[Json.Reader])(using Quotes, Type[Double]): Expr[Double] =
       '{ $reader.double() }
 
   given float: (Float is Inlinable) = new Inlinable:
     type Self = Float
-    def parse(reader: Expr[JsonReader])(using Quotes, Type[Float]): Expr[Float] =
+    def parse(reader: Expr[Json.Reader])(using Quotes, Type[Float]): Expr[Float] =
       '{ $reader.double().toFloat }
 
   given boolean: (Boolean is Inlinable) = new Inlinable:
     type Self = Boolean
-    def parse(reader: Expr[JsonReader])(using Quotes, Type[Boolean]): Expr[Boolean] =
+    def parse(reader: Expr[Json.Reader])(using Quotes, Type[Boolean]): Expr[Boolean] =
       '{ $reader.boolean() }
 
   given text: (Text is Inlinable) = new Inlinable:
     type Self = Text
-    def parse(reader: Expr[JsonReader])(using Quotes, Type[Text]): Expr[Text] =
+    def parse(reader: Expr[Json.Reader])(using Quotes, Type[Text]): Expr[Text] =
       '{ $reader.string() }
 
   given string: (String is Inlinable) = new Inlinable:
     type Self = String
-    def parse(reader: Expr[JsonReader])(using Quotes, Type[String]): Expr[String] =
+    def parse(reader: Expr[Json.Reader])(using Quotes, Type[String]): Expr[String] =
       '{ $reader.string().s }
 
   given iterable: [collection <: Iterable, element]
