+++ lib/stratiform/src/binary/stratiform.BintelReader.scala
@@ -42,7 +42,7 @@ object BintelReader:
   // Only stratiform's read path (`Bintel.parse`) constructs readers, so the
   // exclusivity of the wrapped parser and the resolution scope of the
   // carried tactic are preserved by construction. The wrapped tactic
-  // travels as a neutral carrier (jacinta's `JsonReader` pattern): the
+  // travels as a neutral carrier (jacinta's `Json.Reader` pattern): the
   // field stays pure, and each accessor reasserts the type at the rim —
   // the audited point.
   private[stratiform] def apply(parser: BintelParser, tactic: Tactic[Bintel.Error])
