+++ lib/xenophile/src/kotlin/soundness_xenophile_kotlin.scala
@@ -33,4 +33,4 @@
 package soundness
 
 export xenophile.{companion, enumEntry, make, singleton, Facade, Kotlin, KotlinDialect,
-    KotlinFacade, KotlinInvoke, KotlinRuntime}
+    KotlinFacade, KotlinInvoke}
